Containerizing routes
