+++ app/lib/repositories/textsRepository.tsx
@@ -1,32 +1,4 @@
-export type GitaChapter = {
-  id: string;
-  chapter: number;
-  title: string;
-  author: string;
-  summary: string;
-  commentary: string;
-  analysis: string;
-  verses: Verse[];
-};
-
-export type Shkloka = {
-  id: string;
-  title: string;
-  author: string;
-  summary: string;
-  verses: Verse[];
- }
-
-
-export type Verse = {
-  id: string;
-  verse: number;
-  sanskrit: string;
-  transliteration: string;
-  translation: string;
-  commentary: string;
-};
-
+import type { GitaChapter, Shkloka } from "~/types";
 
 export function getRibhuGitaChapter(): GitaChapter {
