+++ app/routes/articles.$slug.tsx
@@ -1,9 +1,9 @@
 import { type LoaderFunctionArgs } from "@remix-run/node";
-import { useLoaderData, Link } from "@remix-run/react";
-import { marked } from "marked";
+import { useLoaderData } from "@remix-run/react";
+import { ArticleDetailContainer } from "~/components/Articles/ArticleDetailContainer";
 import { PageFrame } from "~/components/Layout/PageFrame";
-import { PageHeader } from "~/components/Layout/PageHeader";
 import { getArticleById } from "~/loader/articles";
+
 
 export async function loader({ params }: LoaderFunctionArgs) {
   const { slug } = params;
@@ -22,22 +22,7 @@ export default function ArticleDetail() {
 
   return (
     <PageFrame>
-      <Link to="/articles" className="text-muted hover:underline text-lg mb-2 inline-block text-gray-500">
-        ← Back
-      </Link>
-      <header className="mb-12">
-        <PageHeader>{article.title}</PageHeader>
-      </header>
-      <div className="max-w-4xl mx-auto">
-        <p className="mt-2 text-lg text-gray-600 italic">{article.summary}</p>
-        <hr className="mt-4 text-sm text-gray-500">
-        </hr>
-        <article className="prose lg:prose-xl">
-          <section className="mt-8 text-gray-800 leading-relaxed">
-            <div className="space-y-6" dangerouslySetInnerHTML={{ __html: marked(article.content) }} />
-          </section>
-        </article>
-      </div>
+      <ArticleDetailContainer article={article} />
     </PageFrame>
   );
 }
