+++ app/routes/texts.ganesha-panchratnam.tsx
@@ -2,9 +2,7 @@ import { type LoaderFunctionArgs } from "@remix-run/node";
 import type { MetaFunction } from "@remix-run/node";
 import { getGaneshaPanchratnamData } from "~/loader/texts";
 import { useLoaderData } from "@remix-run/react";
-import { VerseList } from "~/components/VerseList";
-import { PageFrame } from "~/components/Layout/PageFrame";
-import { PageHeader } from "~/components/Layout/PageHeader";
+import TextDisplayContainer from "~/components/Texts/TextDisplayContainer";
 
 export const meta: MetaFunction = () => {
   return [
@@ -19,13 +17,5 @@ export async function loader({ }: LoaderFunctionArgs) {
 export default function Gita() {
   const { verses, title, summary, author } = useLoaderData<typeof loader>();
 
-
-  return (
-    <PageFrame>
-      <PageHeader>{title}</PageHeader>
-      <h2>{author}</h2>
-      <p className="text-gray-700 text-lg py-4">{summary}</p>
-      <VerseList verses={verses} />
-    </PageFrame>
-  );
+  return <TextDisplayContainer title={title} author={author} summary={summary} verses={verses} />;
 }
