+++ app/routes/texts.nirvana-shaktam.tsx
@@ -2,10 +2,7 @@ import type { MetaFunction } from "@remix-run/node";
 import { type LoaderFunctionArgs } from "@remix-run/node";
 import { useLoaderData } from "@remix-run/react";
 import { getNirvanaShaktamData } from "~/loader/texts";
-import { VerseList } from "~/components/VerseList";
-import { Page } from "openai/pagination.mjs";
-import { PageFrame } from "~/components/Layout/PageFrame";
-import { PageHeader } from "~/components/Layout/PageHeader";
+import TextDisplayContainer from "~/components/Texts/TextDisplayContainer";
 
 export const meta: MetaFunction = () => {
     return [
@@ -21,12 +18,5 @@ export async function loader({ }: LoaderFunctionArgs) {
 export default function NirvanaShaktamPage() {
     const { verses, title, author, summary } = useLoaderData<typeof loader>();
 
-    return (
-        <PageFrame>
-            <PageHeader>{title}</PageHeader>
-            <h2>{author}</h2>
-            <p className="text-gray-700 text-lg py-4">{summary}</p>
-             <VerseList verses={verses} />
-        </PageFrame>
-    );
+    return <TextDisplayContainer title={title} author={author} summary={summary} verses={verses} />;
 }
