+++ package.json
@@ -32,6 +32,7 @@
     "@remix-run/dev": "^2.16.5",
     "@testing-library/react": "^16.3.0",
     "@types/jest": "^29.5.14",
+    "@types/mocha": "^10.0.10",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
     "@types/testing-library__react": "^10.0.1",
